fix: removes unnecessary type restriction
All the queue methods are restricted to a generic type T that should
inherit from ILightMessage.

First, this is the same as asking that the type be an ILightMessage.
However, upon inspecting the method one can see that there's no reason
for the restriction - the methods can work with any objects.